requirements.txt
autopep8==1.6.0
pdf2image==1.16.0
Pillow==8.4.0
pycodestyle==2.8.0
pytesseract==0.3.8
toml==0.10.2
